pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gradient-scraper"
version = "0.1.0"
description = "Scraper for Gradient Academy API"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.1",
    "rich>=13.0.0",
]

[project.scripts]
gradient-scraper = "src.main:main"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100
